Added Login/registration page and worked on the database methods
I added userData annd deleted  channelData. I know the removeChannel, createuser, and checkuser functions are correct but not sure about the other functions

The channel routes is for getting the channel statistics with /:channelid. /delete to delete the channel.
The users routes is for logging in/out and registering

/:channelid shows the channel statistics
Need to make ChannelFound protected- unable to be access if not login